requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-openai
langgraph-cli[inmem]
dotenv-python
asyncio
mcp[cli]
mcp-use